requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
pydantic>=2.5.0
rich>=13.7.0
asyncio-throttle>=1.0.2
flask>=3.0.0
flask-socketio>=5.3.6
numpy>=1.26.0
networkx>=3.2.0
scipy>=1.12.0
